Change return type of pgp_block_size/pgp_key_size to size_t and eliminate coverity issue.

+++ src/lib/crypto/symmetric.h
@@ -130,11 +130,11 @@ typedef struct pgp_crypt_t {
     rnp::RNG *     rng;
 } pgp_crypt_t;
 
-unsigned pgp_block_size(pgp_symm_alg_t);
-unsigned pgp_key_size(pgp_symm_alg_t);
-bool     pgp_is_sa_supported(int alg, bool silent = false);
-bool     pgp_is_sa_aes(pgp_symm_alg_t alg);
-size_t   pgp_cipher_block_size(pgp_crypt_t *crypt);
+size_t pgp_block_size(pgp_symm_alg_t);
+size_t pgp_key_size(pgp_symm_alg_t);
+bool   pgp_is_sa_supported(int alg, bool silent = false);
+bool   pgp_is_sa_aes(pgp_symm_alg_t alg);
+size_t pgp_cipher_block_size(pgp_crypt_t *crypt);
 
 /**
  * Initialize a cipher object.
